move CollectionAppliedHardwareAcc to exitsing common module
