+++ lib/collection/src/common/mod.rs
@@ -2,6 +2,7 @@ pub mod batching;
 pub mod eta_calculator;
 pub mod fetch_vectors;
 pub mod file_utils;
+pub mod hardware_counting;
 pub mod is_ready;
 pub mod retrieve_request_trait;
 pub mod sha_256;
