+++ lib/collection/src/discovery.rs
@@ -1,18 +1,11 @@
 use std::time::Duration;
 
-use futures::Future;
-use itertools::Itertools;
-use segment::data_types::vectors::NamedQuery;
-use segment::types::{Condition, Filter, HasIdCondition, ScoredPoint};
-use segment::vector_storage::query::{ContextPair, ContextQuery, DiscoveryQuery};
-use tokio::sync::RwLockReadGuard;
-
-use crate::collection::common::CollectionAppliedHardwareAcc;
 use crate::collection::Collection;
 use crate::common::batching::batch_requests;
 use crate::common::fetch_vectors::{
     convert_to_vectors, resolve_referenced_vectors_batch, ReferencedVectors,
 };
+use crate::common::hardware_counting::CollectionAppliedHardwareAcc;
 use crate::common::retrieve_request_trait::RetrieveRequest;
 use crate::operations::consistency_params::ReadConsistency;
 use crate::operations::query_enum::QueryEnum;
@@ -21,6 +14,12 @@ use crate::operations::types::{
     CollectionError, CollectionResult, CoreSearchRequest, CoreSearchRequestBatch,
     DiscoverRequestInternal,
 };
+use futures::Future;
+use itertools::Itertools;
+use segment::data_types::vectors::NamedQuery;
+use segment::types::{Condition, Filter, HasIdCondition, ScoredPoint};
+use segment::vector_storage::query::{ContextPair, ContextQuery, DiscoveryQuery};
+use tokio::sync::RwLockReadGuard;
 
 fn discovery_into_core_search(
     collection_name: &str,
