+++ lib/collection/tests/integration/multi_vec_test.rs
@@ -3,8 +3,8 @@ use std::num::NonZeroU32;
 use std::path::Path;
 
 use api::rest::SearchRequestInternal;
-use collection::collection::common::CollectionAppliedHardwareAcc;
 use collection::collection::Collection;
+use collection::common::hardware_counting::CollectionAppliedHardwareAcc;
 use collection::config::{CollectionConfig, CollectionParams, WalConfig};
 use collection::operations::point_ops::{
     PointInsertOperationsInternal, PointOperations, PointStructPersisted, VectorStructPersisted,
